Factor out other-checkbox-form-field.tsx.
